+++ test/plugins/dead_plugin/dead_plugin.py
@@ -1,0 +1,2 @@
+def foobarbar():
+    print("I should not be loaded as I'm not enabeled")